refactor: extract shared build.rs version-injection into beamtalk-build crate (BT-2172)
The four version-injection functions in beamtalk-cli/build.rs and
beamtalk-compiler-port/build.rs were 97% identical copy-pastes. Extract
them into a new beamtalk-build workspace crate so both build scripts
share a single source of truth and future changes only need to happen
once.
