+++ crates/beamtalk-build/Cargo.toml
@@ -1,0 +1,22 @@
+# Copyright 2026 James Casey
+# SPDX-License-Identifier: Apache-2.0
+
+[package]
+name = "beamtalk-build"
+description = "Shared build-script helpers for beamtalk crates (version injection)"
+version.workspace = true
+authors.workspace = true
+edition.workspace = true
+rust-version.workspace = true
+license.workspace = true
+repository.workspace = true
+readme = "README.md"
+keywords = ["build", "beam", "version"]
+categories = ["development-tools"]
+
+[lib]
+name = "beamtalk_build"
+path = "src/lib.rs"
+
+[lints]
+workspace = true
